Jami mini swerve (#10)
* Unfinished, swerve summer config values added.

* Changed AHRS to Pigeon imu and Lamprey to CanCoder

* gyro changed to m_pigeon, auto and commands added

* Updates

* Merging Pigeon an ids

* Added library from team 3512

Team 3512 wrote some helpful objects for swerve stuff.

Sets proper status frames for Sparkmaxs and cancoder.

Prepares a CanCoder configuration for mk4 swerve modules.

Optimizes the steering angle for sparkmaxs (which we are already doing but this cleans it up into an organized and commented object.)

* Config file swerve constants reorganized

* Swerve Teleop updated

* Updates 1

* Swerve Updates

* Small Touchups

* Added swerve test states for 90 deg, and 1.5 speed

* Swerve Updates

* Added SyncSteerEncoders command

Made a command that checks if the neo integrated encoder was properly reset to the cancoder+offset value at bootup. If it was not properly set, it will try to set it every few seconds until it is set properly.

* Added ErrorCheck.java

ErrorCheck has functions that can be imported as static and can catch any REVLibError's or ErrorCode's that return those codes (hardware devices).

* Fix problem caused by BurnFlash

BurnFlash needs to be called after doing passing all of the settings to the sparkmaxs. It was casuing issues when it setting PID values right after calling burnFlash.

* Fix comparing doubles

The range needs to be accurate enough that were are happy with the value if the module happens to be pointing forward when booted up. It's now checking to make sure the cancoder and neo encoder match to within 0.1 degrees.

* Added AutoSelector and Pathweaver paths

* Reduce Speed on Paths

Doing some testing on a zoom call with some software people.

Im reducing the speed so I dont forget later and have it break something...

* Added field2d

* Small PathWeaver Changes

* Error Check in swerve module added

* Tested Auto routines

Changed the paths a bit + small tuning changes + speed up the paths.

* Slew Rate Limiting added to code

* Current Limiter Change

Co-authored-by: ErikCald <erikcaell@gmail.com>
Co-authored-by: Erik <43829793+ErikCald@users.noreply.github.com>