Fixed SyncArmEncoders.java
I had to have 3 seperate sequences of event to sync the encoders.

First state waits until both devices are booted up.
Second state collects a certain number of samples of the absolute encoder.
Third state repeatedly sets the neo encoder with the average of the samples until it gets updated.